PWM input working
PWM inputs are setup for:
Timer15: Ch1, Ch2
Timer 17: Ch1
These channel inputs are serviced using an interrupt that fully parses the rise/fall of pwm inputs.
The duration high can be retrieved using provide_channel function.
The time high is measured in 1mS/1000 so uS. an int is returned with the requested value